requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.14.2
certifi==2025.10.5
click==8.3.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
pydantic==2.5.0
pydantic_core==2.14.1
python-dotenv==1.0.0
PyYAML==6.0.3
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.48.0